refactor: mirror test suite structure
