+++ src/Arius.Core.Tests/Shared/ChunkStorage/ChunkStorageServiceUploadTests.cs
@@ -186,16 +186,3 @@ public class ChunkStorageServiceUploadTests
         metadata.Metadata[BlobMetadataKeys.AriusType].ShouldBe(BlobMetadataKeys.TypeThin);
     }
 }
-
-internal sealed class NonSeekableReadStream(byte[] content) : MemoryStream(content, writable: false)
-{
-    public override bool CanSeek => false;
-
-    public override long Seek(long offset, SeekOrigin loc) => throw new NotSupportedException();
-
-    public override long Position
-    {
-        get => base.Position;
-        set => throw new NotSupportedException();
-    }
-}
