+++ src/Arius.Core.Tests/Shared/FileTree/FileTreeBlobSerializerTests.cs
@@ -2,7 +2,7 @@ using Arius.Core.Shared.Encryption;
 using Arius.Core.Shared.FileTree;
 using Shouldly;
 
-namespace Arius.Core.Tests.FileTree;
+namespace Arius.Core.Tests.Shared.FileTree;
 
 public class FileTreeBlobSerializerTests
 {
