+++ src/Arius.Core.Tests/Shared/FileTree/UnsortedSnapshotBlobContainerService.cs
@@ -1,0 +1,54 @@
+using Arius.Core.Shared.Storage;
+
+namespace Arius.Core.Tests.Shared.FileTree;
+
+internal sealed class UnsortedSnapshotBlobContainerService(IReadOnlyList<string> snapshots) : IBlobContainerService
+{
+    public bool FileTreesListed { get; private set; }
+
+    public Task CreateContainerIfNotExistsAsync(CancellationToken cancellationToken = default) => Task.CompletedTask;
+
+    public Task UploadAsync(string blobName, Stream content, IReadOnlyDictionary<string, string> metadata, BlobTier tier, string? contentType = null, bool overwrite = false, CancellationToken cancellationToken = default) =>
+        throw new NotSupportedException();
+
+    public Task<Stream> OpenWriteAsync(string blobName, string? contentType = null, CancellationToken cancellationToken = default) =>
+        throw new NotSupportedException();
+
+    public Task<Stream> DownloadAsync(string blobName, CancellationToken cancellationToken = default) =>
+        throw new NotSupportedException();
+
+    public Task<BlobMetadata> GetMetadataAsync(string blobName, CancellationToken cancellationToken = default) =>
+        Task.FromResult(new BlobMetadata { Exists = false });
+
+    public async IAsyncEnumerable<string> ListAsync(string prefix, [System.Runtime.CompilerServices.EnumeratorCancellation] CancellationToken cancellationToken = default)
+    {
+        if (prefix == BlobPaths.Snapshots)
+        {
+            foreach (var snapshot in snapshots)
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                yield return snapshot;
+                await Task.Yield();
+            }
+
+            yield break;
+        }
+
+        if (prefix == BlobPaths.FileTrees)
+            FileTreesListed = true;
+
+        yield break;
+    }
+
+    public Task SetMetadataAsync(string blobName, IReadOnlyDictionary<string, string> metadata, CancellationToken cancellationToken = default) =>
+        throw new NotSupportedException();
+
+    public Task SetTierAsync(string blobName, BlobTier tier, CancellationToken cancellationToken = default) =>
+        throw new NotSupportedException();
+
+    public Task CopyAsync(string sourceBlobName, string destinationBlobName, BlobTier destinationTier, RehydratePriority? rehydratePriority = null, CancellationToken cancellationToken = default) =>
+        throw new NotSupportedException();
+
+    public Task DeleteAsync(string blobName, CancellationToken cancellationToken = default) =>
+        throw new NotSupportedException();
+}
