+++ src/Arius.Explorer.Tests/Settings/FakeApplicationSettings.cs
@@ -1,0 +1,13 @@
+using Arius.Explorer.Settings;
+using System.Collections.ObjectModel;
+
+namespace Arius.Explorer.Tests.Settings;
+
+internal sealed class FakeApplicationSettings : IApplicationSettings
+{
+    public ObservableCollection<RepositoryOptions> RecentRepositories { get; } = [];
+    public int RecentLimit { get; set; } = 10;
+    public int SaveCalls { get; private set; }
+
+    public void Save() => SaveCalls++;
+}
